refactor: Encapsulate LoadingScreen logic in a separate saga
Annotate everything that is not reasonably detachable with '@deprecated #3642' to remove in the future
