+++ packages/shared/loading/actions.ts
@@ -1,5 +1,5 @@
-import { LoadableScene } from 'shared/types'
-import { action } from 'typesafe-actions'
+import {LoadableScene} from 'shared/types'
+import {action} from 'typesafe-actions'
 
 export const SCENE_LOAD = '[SCENE MANAGER] Loading scene'
 export const SCENE_START = '[SCENE MANAGER] Started scene'
@@ -7,7 +7,6 @@ export const SCENE_FAIL = '[SCENE MANAGER] Failed scene'
 export const SCENE_UNLOAD = '[SCENE MANAGER] Unload scene'
 export const SCENE_CHANGED = '[SCENE MANAGER] Scenes changed'
 export const PENDING_SCENES = '[SCENE MANAGER] Pending count'
-export const UPDATE_LOADING_SCREEN = '[RENDERER] Refresh loading screen visible'
 
 export const scenesChanged = () => action(SCENE_CHANGED)
 export const signalSceneLoad = (scene: LoadableScene) => action(SCENE_LOAD, scene)
@@ -28,4 +27,3 @@ export const updateStatusMessage = (message: string, loadPercentage: number) =>
   action(UPDATE_STATUS_MESSAGE, { message, loadPercentage })
 export type UpdateStatusMessage = ReturnType<typeof updateStatusMessage>
 
-export const updateLoadingScreen = () => action(UPDATE_LOADING_SCREEN)
