+++ packages/shared/renderer/selectors.ts
@@ -1,12 +1,8 @@
-import { store } from 'shared/store/isolatedStore'
-import { RendererModules, RootRendererState } from './types'
+import {store} from 'shared/store/isolatedStore'
+import {RendererModules, RootRendererState} from './types'
 
 export function isRendererInitialized(state: RootRendererState) {
   return state && state.renderer && state.renderer.initialized
-}
-
-export function getParcelLoadingStarted(state: RootRendererState) {
-  return state && state.renderer && state.renderer.parcelLoadingStarted
 }
 
 export function getClientPort(state: RootRendererState) {
